初始化前端项目

+++ modules/account/account-model/src/main/java/com/lebaoxun/modules/account/entity/UserEntity.java
@@ -25,7 +25,7 @@ public class UserEntity implements Serializable {
 	/**
 	 * 用户ID
 	 */
-	private Integer userId;
+	private Long userId;
 	/**
 	 * 用户账号
 	 */
@@ -146,13 +146,13 @@ public class UserEntity implements Serializable {
 	/**
 	 * 设置：用户ID
 	 */
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 	/**
 	 * 获取：用户ID
 	 */
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 	/**
